"implementado operacoes de save, busca, update e delete de tasks e integracao com o UserMS"

+++ code/taskMS/src/main/java/com/mentoriaprogramacao/taskMS/domain/service/listTasksService/ListTaskService.java
@@ -2,19 +2,17 @@ package com.mentoriaprogramacao.taskMS.domain.service.listTasksService;
 
 import com.mentoriaprogramacao.taskMS.adapter.dto.ListTaskRequest;
 import com.mentoriaprogramacao.taskMS.domain.entity.ListTasksEntity;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 public interface ListTaskService {
     List<ListTasksEntity> findAllLists();
 
-    ListTasksEntity findListsByID(Integer id);
+    ListTasksEntity findListsByID(long id);
 
-    void deleteList(Integer id);
+    void deleteList(long id);
 
-    ListTasksEntity updateList(Integer id, ListTaskRequest request);
+    ListTasksEntity updateList(long id, ListTaskRequest request);
 
     ListTasksEntity saveTaskList(ListTaskRequest request);
 }
